requirements.txt
discord.py==1.7.2
numpy==1.20.2
python-dotenv==0.17.1
requests==2.25.1